proxy-agent/pyproject.toml
[project]
name = "proxy-agent"
version = "1.0.0"
description = "家宽代理池 Agent 端 - 负责管理和监控本地代理服务"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=12.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
proxy-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "main.py"]
